rework concurrent package loading
